requirements.txt
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0rc1
httplib2==0.22.0
idna==3.10
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
